abstract base ADCClient logic to ADCBaseClient
